Remove included `inputs` dir and model files
We will bundle our model files as a separate repository. The code
depended on the relative location of this directory, so I modified the
input arguments to take full paths. Also, some data files not exclusive
to the models were stored in this directory. They have been moved to
src.ptarcade.data , and will be accessed using